v0.6.2 - Update CBIG_fMRI_Preprocessing
Why:

To create a stand-alone repository for CBIG_fMRI_Preprocessing, we
1. Add a script to copy files from CBIG repo to the stand-alone repo, and remove the stable projects not used for preprocessing.
2. Update README.md.

Except for preparation for stand-alone repository, we
3. Add a unit test: check correctness of 419x419 functional connectivity matrices.

This change addresses the need by:

* CBIG fMRI preprocessing code users

+++ external_packages/SD/SDv1.5.1-svn593/BasicTools/MARS_linearInterpolate_kdTree.m
@@ -39,5 +39,5 @@ function [vals, NViF, NF] = MARS_linearInterpolate_kdTree(points, mesh, data)
 %MARS Project (Thomas Yeo and Mert Sabuncu), MIT, CSAIL, (c) 2008
 %Author: Mert
 [nearest_vertex_guess, d] =  MARS_findNV_kdTree(points, mesh.vertices);
-[vals, NViF, NF] = MARS_linearInterpolateAux(single(points), mesh.vertices, mesh.faces, mesh.vertexNbors,... 
-                                      mesh.vertexFaces, nearest_vertex_guess, data);
+[vals, NViF, NF] = MARS_linearInterpolateAux(single(points), single(mesh.vertices), int32(mesh.faces), int32(mesh.vertexNbors),... 
+                                      int32(mesh.vertexFaces), int32(nearest_vertex_guess), single(data));
